Second commit
(forgot to add description for initial commit AKA commit 1.
Commit 1 adds:
Base file
/categories/ with /categories/create
a form to add a new product which includes an add button, a text field, and error messages / success messages using flash. It will redisplay the form if an error happens and if not, it will then add the category to the category database. Redirects to /categories/ where it displays a table of all added categories and their values.

COMMIT 2 CHANGES:
Fixed up the edit and delete buttons on the index (delete wasn't required), they now properly send you to the right page (edit uses PATCH). When edit is clicked, you will be sent to a edit form where the data of the category you selected will be shown. you can edit the name and save it, cancel the edit, etc.... Error checking was added onto this (same way as last) and it will only add to DB if it meets the requirements. Soft delete was not used inside of commit 1.... it is now in commit 2.